perf: convert hero/footer to server components, extract catalog client
- HeroSection: server component with CSS animations (no motion.js in LCP path)
- Footer: server component (no client JS needed)
- CatalogSection: extracted client component for interactive filtering
- page.tsx: server component composing server + client parts
- CSS: add fade-in-up keyframe animation
- Fix contrast: active filter buttons use border style (5:1+ ratio)
- Bump dark text-muted to #8b8b96 (5.9:1 contrast)
