+++ app/components/course/course-filters.tsx
@@ -59,7 +59,7 @@ export function CourseFilters({
             onClick={() => onTrackChange(null)}
             className={`rounded-full px-3 py-1.5 text-sm transition-colors ${
               selectedTrack === null
-                ? "bg-solana-purple/20 text-solana-purple font-medium"
+                ? "border border-solana-purple text-solana-purple font-medium"
                 : "border border-edge text-content-secondary hover:text-content"
             }`}
           >
@@ -71,7 +71,7 @@ export function CourseFilters({
               onClick={() => onTrackChange(track)}
               className={`rounded-full px-3 py-1.5 text-sm transition-colors ${
                 selectedTrack === track
-                  ? "bg-solana-purple/20 text-solana-purple font-medium"
+                  ? "border border-solana-purple text-solana-purple font-medium"
                   : "border border-edge text-content-secondary hover:text-content"
               }`}
             >
