+++ app/components/landing/footer.tsx
@@ -1,22 +1,18 @@
-"use client";
-
-import { useTranslations } from "next-intl";
+import { getTranslations } from "next-intl/server";
 import Link from "next/link";
 
-export function Footer() {
-  const t = useTranslations("footer");
+export async function Footer() {
+  const t = await getTranslations("footer");
 
   return (
     <footer className="border-t border-edge-soft bg-surface-secondary">
       <div className="mx-auto max-w-7xl px-4 py-12">
         <div className="grid gap-8 sm:grid-cols-2 lg:grid-cols-4">
-          {/* Brand */}
           <div>
             <p className="text-sm font-bold text-content">Superteam Academy</p>
             <p className="mt-2 text-xs text-content-muted">{t("tagline")}</p>
           </div>
 
-          {/* Platform */}
           <div>
             <p className="mb-3 text-[10px] font-semibold uppercase tracking-[0.2em] text-content-muted">{t("platform")}</p>
             <nav className="space-y-2">
@@ -26,7 +22,6 @@ export function Footer() {
             </nav>
           </div>
 
-          {/* Resources */}
           <div>
             <p className="mb-3 text-[10px] font-semibold uppercase tracking-[0.2em] text-content-muted">{t("resources")}</p>
             <nav className="space-y-2">
@@ -36,7 +31,6 @@ export function Footer() {
             </nav>
           </div>
 
-          {/* Community */}
           <div>
             <p className="mb-3 text-[10px] font-semibold uppercase tracking-[0.2em] text-content-muted">{t("community")}</p>
             <nav className="space-y-2">
